slightly weaken aniv timeout scale

+++ tests/timeout/useCases/test_calculateTimeoutDurationUseCase.py
@@ -29,7 +29,7 @@ class TestCalculateTimeoutDurationUseCase:
     @pytest.mark.asyncio
     async def test_invoke_withRandomExponentialTimeoutDuration(self):
         timeoutDuration = RandomExponentialTimeoutDuration(
-            scale = float(9),
+            scale = float(10),
             maximumSeconds = 300,
             minimumSeconds = 30,
         )
